user_and_price_fix

+++ tests/UnitTests/TripServiceTests.cs
@@ -43,7 +43,7 @@ public class TripServiceTests
                 To = "Cracow",
                 BetweenPoints = new List<string> { " Radom " }
             },
-            Price = 50.5f,
+            Price = 50.5m, // ZMIANA: f na m
             Date = DateTime.UtcNow.AddDays(1),
             MaxPassengers = 3
         };
@@ -72,15 +72,15 @@ public class TripServiceTests
     [InlineData("", "Cracow", "Route origin and destination are required.")]
     [InlineData("Warsaw", "", "Route origin and destination are required.")]
     [InlineData("Warsaw", "Warsaw", "Route origin and destination cannot be the same.")]
-    [InlineData("Warsaw", "Cracow", "Trip price must be greater than zero.", -1f)]
-    [InlineData("Warsaw", "Cracow", "Max passengers must be greater than zero.", 50f, 0)]
-    public async Task CreateTrip_InvalidInput_ThrowsValidationException(string from, string to, string expectedMessage, float price = 50f, int maxPassengers = 3)
+    [InlineData("Warsaw", "Cracow", "Trip price must be greater than zero.", -1.0)] // ZMIANA: double
+    [InlineData("Warsaw", "Cracow", "Max passengers must be greater than zero.", 50.0, 0)] // ZMIANA: double
+    public async Task CreateTrip_InvalidInput_ThrowsValidationException(string from, string to, string expectedMessage, double price = 50.0, int maxPassengers = 3) // ZMIANA: double
     {
         // Arrange
         var dto = new CreateTripDTO
         {
             Route = new CreateRouteDTO { From = from, To = to },
-            Price = price,
+            Price = (decimal)price, // ZMIANA: rzutowanie na decimal
             Date = DateTime.UtcNow.AddDays(1),
             MaxPassengers = maxPassengers
         };
@@ -99,7 +99,7 @@ public class TripServiceTests
         var dto = new CreateTripDTO
         {
             Route = new CreateRouteDTO { From = "Warsaw", To = "Cracow" },
-            Price = 50f,
+            Price = 50m, // ZMIANA: f na m
             Date = DateTime.UtcNow.AddDays(-1),
             MaxPassengers = 3
         };
@@ -119,7 +119,7 @@ public class TripServiceTests
         var routeId = Guid.NewGuid();
         var trips = new List<Trip>
         {
-            new Trip { Id = Guid.NewGuid(), DriverId = Guid.NewGuid(), RouteId = routeId, Price = 50f, Date = DateTime.UtcNow.AddDays(1), MaxPassengers = 3, OfferStatus = TripStatus.Active }
+            new Trip { Id = Guid.NewGuid(), DriverId = Guid.NewGuid(), RouteId = routeId, Price = 50m, Date = DateTime.UtcNow.AddDays(1), MaxPassengers = 3, OfferStatus = TripStatus.Active } // ZMIANA: f na m
         };
         var route = new Route { Id = routeId, From = "Warsaw", To = "Cracow" };
 
@@ -141,7 +141,7 @@ public class TripServiceTests
         // Arrange
         var tripId = Guid.NewGuid();
         var routeId = Guid.NewGuid();
-        var trip = new Trip { Id = tripId, DriverId = Guid.NewGuid(), RouteId = routeId, Price = 50f, Date = DateTime.UtcNow.AddDays(1), MaxPassengers = 3, OfferStatus = TripStatus.Active };
+        var trip = new Trip { Id = tripId, DriverId = Guid.NewGuid(), RouteId = routeId, Price = 50m, Date = DateTime.UtcNow.AddDays(1), MaxPassengers = 3, OfferStatus = TripStatus.Active }; // ZMIANA: f na m
         var route = new Route { Id = routeId, From = "Warsaw", To = "Cracow" };
 
         _tripRepositoryMock.Setup(t => t.GetById(tripId)).ReturnsAsync(trip);
